# added endpoint to get performance figures for status transitions