## Context

Added endpoint to return performance statistics for various status transitions: 

Example URL:

https://accredited-programmes-api-dev.hmpps.service.justice.gov.uk/statistics/performance/status-duration?startDate=2022-01-01&statuses=REFERRAL_SUBMITTED,AWAITING_ASSESSMENT

{
    "performance": [
        {
            "status": "AWAITING_ASSESSMENT",
            "averageDuration": "41 hours 30 minutes 57 seconds",
            "averageDurationUnderOneHour": "00 hours 01 minutes 21 seconds",
            "minDuration": "00 hours 00 minutes 01 seconds",
            "maxDuration": "1677 hours 29 minutes 30 seconds"
        },
        {
            "status": "REFERRAL_SUBMITTED",
            "averageDuration": "87 hours 47 minutes 08 seconds",
            "averageDurationUnderOneHour": "00 hours 01 minutes 01 seconds",
            "minDuration": "00 hours 00 minutes 01 seconds",
            "maxDuration": "1846 hours 40 minutes 04 seconds"
        }
    ]
}

<!-- Is there a Trello or Jira ticket you can link to? -->
[APG-216](https://dsdmoj.atlassian.net/browse/APG-216)
<!-- Do you need to add any environment variables? -->
<!-- Is an ADR required? An ADR should be added if this PR introduces a change to the architecture. -->

## Changes in this PR

## Release checklist

[Release process documentation](../doc/how-to/perform-a-release.md)

As part of our continuous deployment strategy we must ensure that this work is
ready to be released once merged.

### Pre-merge

- [ ] There are changes required to the Accredited Programmes UI for this change to work...
  - [ ] ... and they been released to production already

### Post-merge

- [ ] [Manually approve](../doc/how-to/perform-a-release.md#releasing-to-the-preprod-environment) release to preprod
- [ ] [Manually approve](../doc/how-to/perform-a-release.md#releasing-to-the-production-environment) release to prod

<!-- Should a release fail at any step, you as the author should now lead the work to
fix it as soon as possible. You can monitor deployment failures in CircleCI
itself and application errors are found in
[Sentry](https://ministryofjustice.sentry.io/projects/hmpps-accredited-programmes-api/?project=4505330122686464&referrer=sidebar&statsPeriod=24h). -->


[APG-216]: https://dsdmoj.atlassian.net/browse/APG-216?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ